Made word count Sorted by their frequency

+++ MyMapper.java
@@ -3,12 +3,12 @@ package com.Hadoop;
 import java.io.IOException;
 import java.util.StringTokenizer;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
-public class MyMapper extends Mapper<Object, Text, Text, IntWritable> {
-	private final static IntWritable one = new IntWritable(1);
+public class MyMapper extends Mapper<Object, Text, Text, LongWritable> {
+	private final static LongWritable one = new LongWritable(1);
     private Text word = new Text();
       
     public void map(Object key, Text value, Context context ) throws IOException, InterruptedException {
@@ -20,3 +20,4 @@ public class MyMapper extends Mapper<Object, Text, Text, IntWritable> {
     }
 
 }
+
